refactor: interfaces layer 구조 수정
