+++ src/main/java/kr/hhplus/be/server/config/swagger/SwaggerConfig.java
@@ -7,7 +7,7 @@ import io.swagger.v3.oas.models.media.Content;
 import io.swagger.v3.oas.models.media.MediaType;
 import io.swagger.v3.oas.models.responses.ApiResponse;
 import io.swagger.v3.oas.models.responses.ApiResponses;
-import kr.hhplus.be.server.interfaces.common.ErrorResponse;
+import kr.hhplus.be.server.interfaces.api.common.ErrorResponse;
 import org.springdoc.core.customizers.OpenApiCustomizer;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
