+++ src/main/java/kr/hhplus/be/server/interfaces/api/product/ProductController.java
@@ -9,7 +9,7 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import java.util.List;
 import kr.hhplus.be.server.domain.product.ProductService;
 import kr.hhplus.be.server.domain.product.info.ProductInfo;
-import kr.hhplus.be.server.interfaces.common.PagingResponse;
+import kr.hhplus.be.server.interfaces.api.common.PagingResponse;
 import lombok.RequiredArgsConstructor;
 import org.springdoc.core.annotations.ParameterObject;
 import org.springdoc.core.converters.models.PageableAsQueryParam;
